feat: prepare public API for alpha release
- created a `slang_solidity::syntax` module, with sub-modules `nodes`, `parser`, and `visitors`.
- renamed `codegen_parser` to `codegen_syntax` accourdingly.
- renamed `Visitable::visit()` to `Visitable::accept_visitor()`.
- made `Visitable::accept_visitor_with_path()` private, since it accepts a `&mut path`, and we expect/pop nodes in a certain order that external callers can violate.
- renamed `ParserOutput` to `ParseOutput` to match `parse()`.
- changed a few internal helpers/constructors to `pub(crate)` instead.

Tested using `cargo rustdoc`.
