+++ crates/codegen/syntax/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "codegen_parser"
+name = "codegen_syntax"
 version.workspace = true
 rust-version.workspace = true
 edition.workspace = true
