+++ crates/codegen/syntax/src/to_scanner_code.rs
