+++ crates/codegen/syntax/src/trie.rs
