+++ crates/codegen/syntax_templates/src/lib.rs
