+++ crates/codegen/syntax_templates/src/shared/language.rs
@@ -3,7 +3,7 @@ pub use std::{collections::BTreeSet, ops::Range, rc::Rc};
 #[allow(deprecated, unused_imports)]
 use semver::Version;
 
-pub use super::{cst, kinds::*, parser_output::ParserOutput};
+pub use super::{cst, kinds::*, parser_output::ParseOutput};
 
 const DEBUG_ERROR_MERGING: bool = false;
 
@@ -62,7 +62,7 @@ impl ParseError {
     }
 }
 
-pub enum ParseResult {
+pub enum ParserResult {
     Pass {
         node: Rc<cst::Node>,
         error: Option<ParseError>,
@@ -87,10 +87,6 @@ impl<'s> Stream<'s> {
             undo_position: 0,
             has_undo: false,
         }
-    }
-
-    pub fn source(&self) -> &'s str {
-        self.source
     }
 
     pub fn position(&self) -> usize {
@@ -190,14 +186,14 @@ fn call_scanner<L, F>(
     scanner: F,
     kind: TokenKind,
     error_message: &str,
-) -> Option<ParserOutput>
+) -> Option<ParseOutput>
 where
     F: Fn(&L, &mut Stream) -> bool,
 {
     let mut stream = Stream::new(input);
     Some(
         if scanner(language, &mut stream) && stream.peek().is_none() {
-            ParserOutput {
+            ParseOutput {
                 parse_tree: Some(cst::Node::token(
                     kind,
                     Range {
@@ -210,7 +206,7 @@ where
                 errors: vec![],
             }
         } else {
-            ParserOutput {
+            ParseOutput {
                 parse_tree: None,
                 errors: vec![ParseError::new(stream.position(), error_message)],
             }
@@ -225,14 +221,14 @@ fn try_call_scanner<L, F>(
     scanner: F,
     kind: TokenKind,
     error_message: &str,
-) -> Option<ParserOutput>
+) -> Option<ParseOutput>
 where
     F: Fn(&L, &mut Stream) -> Option<bool>,
 {
     let mut stream = Stream::new(input);
     scanner(language, &mut stream).map(|result| {
         if result && stream.peek().is_none() {
-            ParserOutput {
+            ParseOutput {
                 parse_tree: Some(cst::Node::token(
                     kind,
                     Range {
@@ -245,7 +241,7 @@ where
                 errors: vec![],
             }
         } else {
-            ParserOutput {
+            ParseOutput {
                 parse_tree: None,
                 errors: vec![ParseError::new(stream.position(), error_message)],
             }
@@ -254,21 +250,21 @@ where
 }
 
 #[allow(dead_code)]
-fn call_parser<L, F>(language: &L, input: &str, parser: F) -> Option<ParserOutput>
+fn call_parser<L, F>(language: &L, input: &str, parser: F) -> Option<ParseOutput>
 where
-    F: Fn(&L, &mut Stream) -> ParseResult,
+    F: Fn(&L, &mut Stream) -> ParserResult,
 {
     let mut stream = Stream::new(input);
     Some(match parser(language, &mut stream) {
-        ParseResult::Pass { node, .. } if stream.peek().is_none() => ParserOutput {
+        ParserResult::Pass { node, .. } if stream.peek().is_none() => ParseOutput {
             parse_tree: Some(node),
             errors: vec![],
         },
-        ParseResult::Pass { .. } => ParserOutput {
+        ParserResult::Pass { .. } => ParseOutput {
             parse_tree: None,
             errors: vec![ParseError::new(stream.position(), "end of input")],
         },
-        ParseResult::Fail { error } => ParserOutput {
+        ParserResult::Fail { error } => ParseOutput {
             parse_tree: None,
             errors: vec![error],
         },
@@ -276,21 +272,21 @@ where
 }
 
 #[allow(dead_code)]
-fn try_call_parser<L, F>(language: &L, input: &str, parser: F) -> Option<ParserOutput>
+fn try_call_parser<L, F>(language: &L, input: &str, parser: F) -> Option<ParseOutput>
 where
-    F: Fn(&L, &mut Stream) -> Option<ParseResult>,
+    F: Fn(&L, &mut Stream) -> Option<ParserResult>,
 {
     let mut stream = Stream::new(input);
     parser(language, &mut stream).map(|result| match result {
-        ParseResult::Pass { node, .. } if stream.peek().is_none() => ParserOutput {
+        ParserResult::Pass { node, .. } if stream.peek().is_none() => ParseOutput {
             parse_tree: Some(node),
             errors: vec![],
         },
-        ParseResult::Pass { .. } => ParserOutput {
+        ParserResult::Pass { .. } => ParseOutput {
             parse_tree: None,
             errors: vec![ParseError::new(stream.position(), "end of input")],
         },
-        ParseResult::Fail { error } => ParserOutput {
+        ParserResult::Fail { error } => ParseOutput {
             parse_tree: None,
             errors: vec![error],
         },
