+++ crates/codegen/syntax_templates/src/shared/macros.rs
