+++ crates/codegen/syntax_templates/src/typescript/cst_types.rs
