+++ crates/codegen/syntax_templates/src/typescript/mod.rs
