+++ crates/codegen/syntax_templates/src/typescript/parser_output.rs
@@ -2,37 +2,44 @@ use std::rc::Rc;
 
 use super::{
     cst,
+    cst_types::RcNodeExtensions as CSTRcNodeExtensions,
     language::{render_error_report, ParseError},
 };
+use napi::bindgen_prelude::*;
 
-#[derive(PartialEq)]
-pub struct ParserOutput {
+#[napi]
+pub struct ParseOutput {
     pub(crate) parse_tree: Option<Rc<cst::Node>>,
     pub(crate) errors: Vec<ParseError>,
 }
 
-impl ParserOutput {
-    pub fn parse_tree(&self) -> Option<Rc<cst::Node>> {
-        self.parse_tree.clone()
+#[napi]
+impl ParseOutput {
+    #[napi(ts_return_type = "RuleNode | null")]
+    pub fn parse_tree(&self, env: Env) -> Option<napi::JsObject> {
+        self.parse_tree.clone().map(|n| n.to_js(&env))
     }
 
+    #[napi]
     pub fn error_count(&self) -> usize {
         self.errors.len()
     }
 
+    #[napi]
     pub fn errors_as_strings(
         &self,
-        source_id: &str,
-        source: &str,
+        source_id: String,
+        source: String,
         with_colour: bool,
     ) -> Vec<String> {
         return self
             .errors
             .iter()
-            .map(|error| render_error_report(error, source_id, source, with_colour))
+            .map(|error| render_error_report(error, &source_id, &source, with_colour))
             .collect();
     }
 
+    #[napi]
     pub fn is_valid(&self) -> bool {
         self.parse_tree.is_some() && self.errors.is_empty()
     }
