+++ crates/solidity/outputs/rust/slang_solidity/src/syntax/mod.rs
@@ -1,0 +1,16 @@
+mod generated;
+
+pub mod nodes {
+    pub use super::generated::cst::Node;
+    pub use super::generated::kinds::{RuleKind, TokenKind};
+}
+
+pub mod parser {
+    pub use super::generated::language::{Language, ParseOutput, ProductionKind};
+}
+
+pub mod visitors {
+    pub use super::generated::cst_visitor::{
+        Visitable, Visitor, VisitorEntryResponse, VisitorExitResponse,
+    };
+}
