+++ crates/solidity/outputs/typescript/lib/Cargo.toml
@@ -18,7 +18,7 @@ serde = { workspace = true }
 
 [build-dependencies]
 anyhow = { workspace = true }
-codegen_parser = { workspace = true }
+codegen_syntax = { workspace = true }
 codegen_schema = { workspace = true }
 codegen_utils = { workspace = true }
 napi-build = { workspace = true }
